+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Support `jcs-eddsa-2022` identity proofs.
 - Support fetching `replies` collection when it is not embedded.
+- Fetch and parse proposals attached to remote actors.
 
 ### Changed